+++ nzmath/bigrandom.py
@@ -1,5 +1,6 @@
 #bigrandom.py
 
+from builtins import range
 import sys
 import random as _random
 import nzmath.arith1 as arith1
@@ -52,11 +53,11 @@ def _validate_for_randrange(start, stop, step):
     """
     if step == 0:
         raise ValueError("zero step for randrange()")
-    elif start != long(start):
+    elif start != int(start):
         raise ValueError("non-integer arg 1 for randrange()")
-    elif stop != long(stop):
+    elif stop != int(stop):
         raise ValueError("non-integer stop for randrange()")
-    elif step != long(step):
+    elif step != int(step):
         raise ValueError("non-integer step for randrange()")
     if start >= stop:
         raise ValueError("empty range for randrange()")
